graphdoc-dep/pyproject.toml
[tool.poetry]
name = "graphdoc-dep"
version = "0.1.0"
description = ""
authors = ["denver <denver@semiotic.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
jupyter = "1.1.1"
ipykernel = "6.29.5"
graphql-core = "3.2.5"
openai = "1.58.1"
jinja2 = "3.1.4"
python-dotenv = "1.0.1"
tokencost = "0.1.17"
subgrounds = "1.9.1"
dspy = {version = "^2.6.0rc8", allow-prereleases = true}


[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-asyncio = "0.25.0"
black = "24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
